requirements.txt

polars>=0.20.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.0.0
torch>=2.2.0
websockets>=12.0
aiohttp>=3.9.0
pyarrow>=15.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
streamlit>=1.30.0
binance-connector>=3.0.0
